+++ lib/Elastica/Transport/AwsAuthV4.php
@@ -6,7 +6,6 @@ use Aws\Credentials\CredentialProvider;
 use Aws\Credentials\Credentials;
 use Aws\Signature\SignatureV4;
 use Elastica\Connection;
-use Elastica\Request;
 use GuzzleHttp;
 use GuzzleHttp\Client;
 use GuzzleHttp\HandlerStack;
@@ -15,18 +14,14 @@ use Psr\Http\Message\RequestInterface;
 
 class AwsAuthV4 extends Guzzle
 {
-    protected function _getGuzzleClient($baseUrl, $persistent = true, Request $request)
+    protected function _getGuzzleClient($persistent = true)
     {
         if (!$persistent || !self::$_guzzleClientConnection) {
             $stack = HandlerStack::create(GuzzleHttp\choose_handler());
             $stack->push($this->getSigningMiddleware(), 'sign');
 
             self::$_guzzleClientConnection = new Client([
-                'base_uri' => $baseUrl,
                 'handler' => $stack,
-                'headers' => [
-                    'Content-Type' => $request->getContentType(),
-                ],
             ]);
         }